Compiles with CTRE

+++ src/parser/material_params.h
@@ -4,9 +4,9 @@
 #include "geometry/spectrum.h"
 
 struct MaterialParameters {
-  float ior;        // index of refraction
-  float cone_angle; // reflection cone angle
+  double ior;        // index of refraction
+  double cone_angle; // reflection cone angle
   Spectrum diffuse;
   Spectrum emissive;
-  float reflectivity;
+  double reflectivity;
 };
